submit F.cpp to 'F - Takahashi's Basics in Education and Learning' (abc129) [C++14 (GCC 5.4.1)]

+++ 2019/0611_ABC129/F.cpp
@@ -368,7 +368,7 @@ int main()
   v = {0, A, 1};
   for (auto i = 1; i <= 18; i++)
   {
-    int x = ub[i] - ub[i - 1];
+    ll x = ub[i] - ub[i - 1];
     if (x == 0)
     {
       continue;
